+++ src/Exception/ConnectException.php
@@ -2,15 +2,12 @@
 
 namespace GuzzleHttp\Exception;
 
-use Psr\Http\Client\NetworkExceptionInterface;
 use Psr\Http\Message\RequestInterface;
 
 /**
  * Exception thrown when a connection cannot be established.
- *
- * Note that no response is present for a ConnectException
  */
-class ConnectException extends TransferException implements NetworkExceptionInterface
+class ConnectException extends NetworkException
 {
     /**
      * @var RequestInterface