+++ tests/Exception/ConnectExceptionTest.php
@@ -3,6 +3,7 @@
 namespace GuzzleHttp\Tests\Exception;
 
 use GuzzleHttp\Exception\ConnectException;
+use GuzzleHttp\Exception\NetworkException;
 use GuzzleHttp\Psr7\Request;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\NetworkExceptionInterface;
@@ -18,6 +19,7 @@ class ConnectExceptionTest extends TestCase
         $req = new Request('GET', '/');
         $prev = new \Exception();
         $e = new ConnectException('foo', $req, $prev, ['foo' => 'bar']);
+        self::assertInstanceOf(NetworkException::class, $e);
         self::assertInstanceOf(NetworkExceptionInterface::class, $e);
         self::assertNotInstanceOf(RequestExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());